created new internal sorting module
